requirements.txt
solana
solders
django
django-solo
django-environ
python-telegram-bot
base58